Add a time output option to bistre.
This is the only bistre output feature I have been missing.
Pass --time to bistre to see a logline's timestamp in gray.
Passing no option retains the current behavior.

Added an output test for it in the same mode as the existing tests.
While I was there, bumped the versions for the chalk & through deps.